requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloud-sql-python-connector==1.15.0
comm==0.2.2
cryptography==44.0.0
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
frozenlist==1.5.0
google-auth==2.37.0
google-cloud==0.34.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
macholib==1.16.3
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pg8000==1.31.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
scramp==1.4.5
screeninfo==0.8.1
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
